requirements.txt
wsgiref==0.1.2
django-mediagenerator
django-registration
Django==1.3
psycopg2==2.4.2
django-tastypie
